// src/app/interfaces/interface-xlsx-work-sheet.ts
export interface InterfaceXlsxWorkSheet {
    country: string;
    segment: string;
    idFinancialGroup: string;
    financialGroup: string;
    totalGroupSantanderExposure: string;
    currencyOfGroupSantanderExposure: string;
    idClient: string;
    client: string;
    economicSector: string;
    feveScan: string;
    ratingClient: string;
    pdAssociatedToRating: string;
    companySantanderExposure: string;
    currencyCompanySantanderExposure: string;
    financialSponsor: string;
    dealIdNumber: string;
    dealAmount: string;
    currencyOfDeal: string;
    originaTionDate: string;
    maturityDate: string;
    product: string;
    typeOfDeal: string;
    useOfFounds: string;
    seniorityBreakDown: string;
    provisionAssociatedToDeal: string;
    ifrs9SageAssociatedToDeal: string;
    covenantTypeAssociatedToDeal: string;
    guarantee: string;
    typeOfGuarantee: string;
    turnOver: string;
    grossFinancialDebtBalance: string;
    cashBalance: string;
    ebitda: string;
    ratioGfdEbitda: string;
    shareOfRisk: string;
    dalancaDate: string;
    currencyBalanceInformation: string;
}